requirements.txt
# Core PyQt6 GUI
PyQt6>=6.5.0
PyQt6-Qt6>=6.5.0

# System monitoring (lightweight)
psutil>=5.9.0

# Process automation (traditional)
pywinauto>=0.6.8
pyautogui>=0.9.54

# OCR for logout detection
pytesseract>=0.3.10
Pillow>=9.0.0
opencv-python-headless>=4.5.0

# Windows API
pywin32>=306

# Remote control
Flask>=2.3.0
requests>=2.28.0

# Configuration and logging
numpy>=1.24.0

# Optional: Telegram notifications
python-telegram-bot>=20.0

# End of requirements.